[feature] Load fixtures for tests in separate file
